+++ extensions/gmail/CHANGELOG.md
@@ -7,5 +7,6 @@
 - Add `List Trash` command
 - `List Mails` now have a optional query parameter
 - `List Mails` command can now store the current query as Quicklink
+- Add actions to select multiple mails
 
 ## [Initial Version] - 2023-08-09